requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.2
Flask-Login==0.6.2
numpy==2.2.3
opencv-python==4.12.0.88
Pillow==10.1.0
face-recognition==1.3.0
